# swagger.yaml
openapi: 3.0.0
info:
  title: 'ProviEmplea API'
  description: 'API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080/api'
    description: 'Servidor de desarrollo local'
paths:
...
  /personas:
    get:
      tags:
        - Personas
      summary: 'Listar personas (CV ciego)'
      description: 'Obtiene talentos activos en formato de CV ciego (sin datos personales identificables).'
      operationId: getPersonas
      parameters:
        -
          name: validado
          in: query
          description: 'Filtrar por validación'
          required: false
          schema:
            type: boolean
        -
          name: nivel_educacional
          in: query
          required: false
          schema:
            type: string
            enum:
              - basica
              - media
              - tecnica
              - universitaria
              - postgrado
      responses:
        '200':
          description: 'Listado exitoso'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonaCVCiego'
    post:
      tags:
        - Personas
      summary: 'Registrar nueva persona/talento'
      description: 'Crea un perfil de talento. El código se genera automáticamente.'
      operationId: createPersona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaInput'
      responses:
        '201':
          description: 'Persona creada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Persona'
        '422':
          description: 'Errores de validación'
  '/personas/{id}':
    get:
      tags:
        - Personas
      summary: 'Obtener persona por ID'
      operationId: getPersona
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Persona encontrada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Persona'
        '404':
          description: 'No encontrada'
    put:
      tags:
        - Personas
      summary: 'Actualizar persona'
      operationId: updatePersona
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaInput'
      responses:
        '200':
          description: 'Persona actualizada'
        '404':
          description: 'No encontrada'
    delete:
      tags:
        - Personas
      summary: 'Desactivar persona'
      description: 'Desactiva el perfil sin eliminarlo de la base de datos.'
      operationId: deletePersona
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Persona desactivada'
        '404':
          description: 'No encontrada'
  '/personas/{id}/validar':
    patch:
      tags:
        - Personas
      summary: 'Validar persona (solo administración)'
      description: 'Marca a una persona como validada para que aparezca en la vitrina.'
      operationId: validarPersona
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Persona validada'
        '404':
          description: 'No encontrada'
components:
  schemas:
...
    PersonaCVCiego:
      title: 'Persona CV Ciego'
      description: 'Vista pública del talento sin datos personales identificables'
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        codigo_talento:
          type: string
          example: PROV-2026-A1B2
        resumen:
          type: string
          nullable: true
        nivel_educacional:
          type: string
          nullable: true
        titulo_carrera:
          type: string
          nullable: true
        anio_egreso:
          type: integer
          nullable: true
        anios_experiencia:
          type: integer
          example: 5
        tipo_jornada:
          type: string
          nullable: true
        modalidad:
          type: string
          nullable: true
        persona_discapacidad:
          type: boolean
          example: false
      type: object
    PersonaInput:
      title: 'Persona Input'
      description: 'Datos para crear o actualizar una persona'
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: juan@example.com
        telefono:
          type: string
          example: '+56912345678'
          nullable: true
        resumen:
          type: string
          nullable: true
        nivel_educacional:
          type: string
          nullable: true
          enum:
            - basica
            - media
            - tecnica
            - universitaria
            - postgrado
        titulo_carrera:
          type: string
          nullable: true
        anio_egreso:
          type: integer
          example: 2020
          nullable: true
        anios_experiencia:
          type: integer
          example: 3
          nullable: true
        areas_experiencia:
          type: array
          items:
            type: string
          example:
            - 'Desarrollo Web'
          nullable: true
        competencias:
          type: array
          items:
            type: string
          example:
            - PHP
            - Laravel
          nullable: true
        rango_renta:
          type: string
          example: 500k-800k
          nullable: true
        tipo_jornada:
          type: string
          nullable: true
          enum:
            - completa
            - part-time
            - por-horas
        modalidad:
          type: string
          nullable: true
          enum:
            - presencial
            - remoto
            - hibrido
        persona_discapacidad:
          type: boolean
          example: false
          nullable: true
        portafolio_url:
          type: string
          format: url
          nullable: true
      type: object
    Persona:
      title: Persona
      description: 'Modelo completo de persona/talento'
      required:
        - email
        - codigo_talento
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: juan@example.com
        telefono:
          type: string
          example: '+56912345678'
          nullable: true
        codigo_talento:
          type: string
          example: PROV-2026-A1B2
        resumen:
          type: string
          nullable: true
        nivel_educacional:
          type: string
          nullable: true
          enum:
            - basica
            - media
            - tecnica
            - universitaria
            - postgrado
        titulo_carrera:
          type: string
          example: 'Ingeniero en Informática'
          nullable: true
        anio_egreso:
          type: integer
          example: 2019
          nullable: true
        anios_experiencia:
          type: integer
          example: 5
        areas_experiencia:
          type: array
          items:
            type: string
          example:
            - 'Desarrollo Web'
            - 'APIs REST'
          nullable: true
        competencias:
          type: array
          items:
            type: string
          example:
            - PHP
            - Laravel
            - MySQL
          nullable: true
        rango_renta:
          type: string
          example: 800k-1.2M
          nullable: true
        tipo_jornada:
          type: string
          nullable: true
          enum:
            - completa
            - part-time
            - por-horas
        modalidad:
          type: string
          nullable: true
          enum:
            - presencial
            - remoto
            - hibrido
        cursos:
          type: array
          items:
            properties:
              nombre:
                type: string
              institucion:
                type: string
              anio:
                type: integer
            type: object
          nullable: true
        idiomas:
          type: array
          items:
            properties:
              idioma:
                type: string
              nivel:
                type: string
                enum:
                  - basico
                  - intermedio
                  - avanzado
                  - nativo
            type: object
          nullable: true
        portafolio_url:
          type: string
          format: url
          nullable: true
        persona_discapacidad:
          type: boolean
          example: false
        validado:
          type: boolean
          example: false
        activo:
          type: boolean
          example: true
        porcentaje_completitud:
          type: integer
          example: 85
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...